backend/requirements.txt
flask
flask-cors
cassandra-driver
bcrypt
pyjwt
